move date time function to idUtils and update click tracking date time to use the same method
