+++ packages/mwp-tracking-plugin/src/util/clickParser.js
@@ -1,3 +1,5 @@
+import { getZonedDateTimeStringWithUTCOffset } from './idUtils';
+
 const DATA_ATTR = 'clicktrack';
 
 /*
@@ -120,9 +122,8 @@ function getTrackClick() {
 			const data = _getData(e);
 
 			// 2. Create click action with metadata
-			const now = new Date(); // TODO: make this a America/New_York timestamp
 			return {
-				timestamp: now.toISOString(),
+				timestamp: getZonedDateTimeStringWithUTCOffset(),
 				lineage: clickLineage.join('<'),
 				linkText: linkText,
 				coords: [x, y],
